Fixes issue #1497 within the apt module
This change is in response to issue #1497 where the apt module would not
properly updating the apt cache in some situations and never returned a state
change on cache update when the module was used without or without an item to
be installed or upgraded.

The change simply allows the apt module to update the cache when
`update_cache` option is used without or without a set `cache_valid_time`. If
`cache_valid_time` is set and the on disk mtime for apt cache is ">" the
provided amount of seconds, which now has a default of 0, the apt cache will
be updated. Additionally if no upgrade, package, or deb is installed or
changed but the apt cache is updated the module will return a changed state
which will help users to know that the state of the environment has changed
due to a task operation, even if it was only an apt cahce update.

Signed-off By: Kevin Carter <kevin.carter@rackspace.com>