Fix the apt update process
This change is in response to issue #1497 where the apt module
would not properly update the apt cache in some situations and
never returned a state change on cache update when the module
was used without an item to be installed or updated.

The change greatly simply updates the cache when `update_cache`
is used without a `cache_valid_time` or when the cache mtime is
> the provided `cache_valid_time`. Additionally if no package,
upgrade, or deb is provided and the apt cache is updated the
module will now return a change state as *True* which will help
users in knwoning that the module functioned as expected.

Signed-off By: Kevin Carter <kevin.carter@rackspace.com>