Fix the SystemD check function
This change refactors the "check_systemd" function such that it now will
attempt to read "/proc/1/comm" first and should the value of that output
be "systemd" the function will return True. If the output of
"/proc/1/comm" is "init" the function will return False. If
"/proc/1/comm" does not exist the function will look for several canary
files, ["/run/systemd/system/", "/dev/.run/systemd/", "/dev/.systemd/"]
and return True if found. If neither "/proc/1/comm" reads "systemd",
does not exist, and none of the canary files are found the function will
return False.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>